jarvis-cyber/requirements.txt
# Dependências do JARVIS Cyber Security System

# Servidor e API
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Autenticação e Segurança
PyJWT>=2.8.0
cryptography>=41.0.7
passlib>=1.7.4
bcrypt>=4.1.2

# Interface Gráfica
PyQt5>=5.15.10
# Alternativa: PySide2>=5.15.8

# Scanner de Rede
python-nmap>=0.7.1
ipaddress>=1.0.23

# HTTP Client
httpx>=0.25.2
requests>=2.31.0

# Base de Dados
sqlite3  # Built-in Python
sqlalchemy>=2.0.23

# Logs e Configuração
pyyaml>=6.0.1
python-dotenv>=1.0.0
colorama>=0.4.6

# Data e Tempo
python-dateutil>=2.8.2

# Processamento JSON
orjson>=3.9.10

# Sistema e OS
psutil>=5.9.6
python-systemd>=234  # Linux only

# IA e Processamento (Opcional)
openai>=1.3.7
transformers>=4.36.0
torch>=2.1.1  # Para modelos locais

# Voz (Opcional)
pyttsx3>=2.90
SpeechRecognition>=3.10.0

# Docker (Para laboratório)
docker>=6.1.3

# Desenvolvimento e Testes
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# Monitoramento
prometheus-client>=0.19.0

# Utilitários
rich>=13.7.0  # Para output colorido
click>=8.1.7   # Para CLI